+++ src/core/public/workspace/index.ts
@@ -3,4 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { WorkspacesStart, WorkspacesService, WorkspacesSetup } from './workspaces_service';
+export {
+  WorkspacesStart,
+  WorkspacesService,
+  WorkspacesSetup,
+  WorkspaceObject,
+} from './workspaces_service';