+++ src/core/server/saved_objects/import/import_saved_objects.ts
@@ -119,6 +119,7 @@ export async function importSavedObjectsFromStream({
         objects: checkConflictsResult.filteredObjects,
         ignoreRegularConflicts: overwrite,
         dataSourceId,
+        savedObjectsClient,
       });
       checkOriginConflictsParams.objects = checkConflictsForDataSourceResult.filteredObjects;
     }