+++ src/plugins/workspace/public/application.tsx
@@ -5,8 +5,9 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { AppMountParameters } from '../../../core/public';
+import { AppMountParameters, ScopedHistory } from '../../../core/public';
 import { OpenSearchDashboardsContextProvider } from '../../opensearch_dashboards_react/public';
+import { WorkspaceFatalError } from './components/workspace_fatal_error';
 import { WorkspaceCreatorApp } from './components/workspace_creator_app';
 import { Services } from './types';
 
@@ -14,6 +15,21 @@ export const renderCreatorApp = ({ element }: AppMountParameters, services: Serv
   ReactDOM.render(
     <OpenSearchDashboardsContextProvider services={services}>
       <WorkspaceCreatorApp />
+    </OpenSearchDashboardsContextProvider>,
+    element
+  );
+
+  return () => {
+    ReactDOM.unmountComponentAtNode(element);
+  };
+};
+
+export const renderFatalErrorApp = (params: AppMountParameters, services: Services) => {
+  const { element } = params;
+  const history = params.history as ScopedHistory<{ error?: string }>;
+  ReactDOM.render(
+    <OpenSearchDashboardsContextProvider services={services}>
+      <WorkspaceFatalError error={history.location.state?.error} />
     </OpenSearchDashboardsContextProvider>,
     element
   );